Add accrue type

+++ Assets/Mana/BlueMana.cs
@@ -1,14 +1,14 @@
 ﻿public class BlueMana
 {
-    public static void BurnBlue(float amount)
+    public static void BurnBlue(int amount)
     {
     }
 
-    public static void GainBlue(float amount)
+    public static void GainBlue(int amount)
     {
     }
 
-    public static Mana GetBase(float startingTotal = 0)
+    public static Mana GetBase(int startingTotal = 0)
     {
         return new Mana(ManaColor.Blue, GainBlue, BurnBlue)
         {
